[MTDSA-1627] refactor func test packages
