+++ conf/application.conf
@@ -258,7 +258,7 @@ Test {
   }
 
   auditing {
-    enabled = true
+    enabled = false
     traceRequests = false
     consumer {
       baseUri {
